Implement account activation feature
A new feature has been added to handle account activation. It includes generating and storing a unique token for each newly registered user, and checking the token to activate the account. It also involves modifying the user's status depending on if the account is activated or not. Added the necessary classes and interfaces to handle the confirmation tokens and added new columns to the user table in the database for activation status. The Kafka consumer's value deserializer is updated to handle JSON for ActivationEmail object.