Add enhanced profile viewing and edit API security
The main addition in this commit is the enhancement of the profile viewing and editing features in the API. Now, all types of users (admin, manager, client) can get their profile details as per their role with the help of JWT token to ensure security. For editing a user's profile, we check the token to match the username with the requesting user for preventing unauthorized requests. Moreover, updates have been done for clean application of JWT token in filtering requests and handling related exceptions. Several utility methods are added in the user service for finding users and managers. Last but not least, minor clean-ups are done including the schema and data scripts, improving logging using SLF4J, and maintaining standards in entity classes.