Refactor NotificationService, UsersController
Converted DTOs in the notification service to records for more concise syntax and removed unwanted files. Also modified UsersController in the user service to send the entire User object embedded inside ActivationEmail to the Kafka topic. Enhanced logging in the EmailLog class by adding a toString() method for better readability. Additionally, changed an import path due to renaming a package.